+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### v5.0.0-rc.1 (???)
 
-- LCP and CLS attribution now saves the target selector for when the element is removed from the DOM ([#562](https://github.com/GoogleChrome/web-vitals/pull/562))
+- **[BREAKING]** Change LCP selector from element to target. LCP and CLS attribution now saves the target selector for when the element is removed from the DOM ([#562](https://github.com/GoogleChrome/web-vitals/pull/562))
 
 ### v5.0.0-rc.0 (2024-10-03)